+++ CHANGELOG.md
@@ -11,6 +11,15 @@ project adheres to [Semantic Versioning](http://semver.org/).
     HTML files served by the built-in static server (via version bump to
     `koa-node-resolve`).
 
+-   Browsers in the JSON config file can now be specified as an object, e.g.
+    `browser: { name: 'chrome', headless: true }`. The string format is still
+    supported, though more options will be supported by the object form (e.g.
+    `windowSize` below).
+
+-   Added `--window-size` flag and `browser:{ windowSize: {width, height} }`
+    JSON config file property to control browser window size. Browsers will be
+    resized to 1024x768 by default.
+
 <!-- Add new, unreleased changes here. -->
 
 ## [0.4.7] 2019-06-14